pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scheme-advisor-env"
version = "1.0.0"
description = "OpenEnv environment for Indian government welfare scheme eligibility advising"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }

keywords = ["openenv", "reinforcement-learning", "government-schemes", "india", "ai-agents"]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "openenv-core>=0.2.0",
    "requests>=2.25.0",
    "websockets>=11.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
server = "app:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
